refactor: modular structure + CI/CD pipeline (#3, #4)
- Split monolithic lib.rs (~557 lines) into focused modules:
  - errors.rs: ErrorCode enum
  - admin.rs: initialize, upgrade, config getters/setters
  - sources.rs: add/remove/is_source/get_oracle_sources
  - assets.rs: register/unregister/is_asset_registered
  - prices.rs: submit_price, get_price, SEP-40 price queries
  - history.rs: get/has/get_historical_prices
  - lib.rs: thin entry point with mod declarations + delegation
- Add .github/workflows/contract-ci.yml per issue #4 spec
- All 56 tests pass, clippy clean, zero warnings

Closes #3
Closes #4
