+++ contracts/price-oracle/src/lib.rs
@@ -1,6 +1,12 @@
 #![no_std]
 
+mod admin;
+mod assets;
+mod errors;
 mod events;
+mod history;
+mod prices;
+mod sources;
 mod storage;
 mod types;
 
@@ -9,15 +15,9 @@ pub use types::{
     PriceHistoryEntry,
 };
 
-use soroban_sdk::{contract, contractimpl, panic_with_error, Address, Env, String, Symbol, Vec};
+use soroban_sdk::{contract, contractimpl, Address, Env, String, Symbol, Vec};
 
-use crate::events::*;
-use crate::storage::*;
-
-const DEFAULT_MAX_HISTORY: u32 = 100;
-const DEFAULT_MIN_SOURCES: u32 = 1;
-const DEFAULT_DECIMALS: u32 = 18;
-const DEFAULT_RESOLUTION: u32 = 0;
+use crate::storage::read_registered_assets;
 
 #[contract]
 pub struct PriceOracleContract;
@@ -25,6 +25,8 @@ pub struct PriceOracleContract;
 #[contractimpl]
 impl PriceOracleContract {
     pub fn __constructor(_env: Env) {}
+
+    // --- Admin ---
 
     pub fn initialize(
         env: Env,
@@ -34,508 +36,123 @@ impl PriceOracleContract {
         decimals: u32,
         description: String,
     ) {
-        if env.storage().persistent().has(&DataKey::Admin) {
-            panic_with_error!(env, ErrorCode::AlreadyInitialized);
-        }
-        admin.require_auth();
-        env.storage().persistent().set(&DataKey::Admin, &admin);
-        env.storage().persistent().set(
-            &DataKey::MinSourcesRequired,
-            &if min_sources_required > 0 {
-                min_sources_required
-            } else {
-                DEFAULT_MIN_SOURCES
-            },
-        );
-        env.storage().persistent().set(
-            &DataKey::MaxHistoryLength,
-            &if max_history_length > 0 {
-                max_history_length
-            } else {
-                DEFAULT_MAX_HISTORY
-            },
-        );
-        env.storage()
-            .persistent()
-            .set(&DataKey::Resolution, &DEFAULT_RESOLUTION);
-        env.storage()
-            .persistent()
-            .set(&DataKey::Decimals, &decimals);
-        env.storage()
-            .persistent()
-            .set(&DataKey::Description, &description);
-        env.storage().persistent().set(
-            &DataKey::OracleSources,
-            &OracleSources {
-                sources: Vec::new(&env),
-                metadata: soroban_sdk::Map::new(&env),
-            },
-        );
-        env.storage()
-            .persistent()
-            .set(&DataKey::RegisteredAssets, &Vec::<Address>::new(&env));
+        admin::initialize(&env, admin, min_sources_required, max_history_length, decimals, description);
     }
 
     pub fn upgrade(env: Env, new_wasm_hash: soroban_sdk::BytesN<32>) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        ContractUpgradedEvent {
-            new_wasm_hash: new_wasm_hash.clone(),
-        }
-        .publish(&env);
-        env.deployer().update_current_contract_wasm(new_wasm_hash);
+        admin::upgrade(&env, new_wasm_hash);
     }
 
     pub fn set_admin(env: Env, new_admin: Address) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        env.storage().persistent().set(&DataKey::Admin, &new_admin);
-        AdminChangedEvent {
-            new_admin: new_admin.clone(),
-        }
-        .publish(&env);
+        admin::set_admin(&env, new_admin);
     }
 
     pub fn get_admin_address(env: Env) -> Address {
-        if env.storage().persistent().has(&DataKey::Admin) {
-            env.storage()
-                .persistent()
-                .extend_ttl(&DataKey::Admin, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        get_admin(&env)
+        admin::get_admin_address(&env)
     }
 
     pub fn set_min_sources_required(env: Env, new_min: u32) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        env.storage()
-            .persistent()
-            .set(&DataKey::MinSourcesRequired, &new_min);
-        MinSourcesChangedEvent { value: new_min }.publish(&env);
+        admin::set_min_sources_required(&env, new_min);
     }
 
     pub fn get_min_sources_required(env: Env) -> u32 {
-        let key = DataKey::MinSourcesRequired;
-        if env.storage().persistent().has(&key) {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        env.storage()
-            .persistent()
-            .get(&key)
-            .unwrap_or(DEFAULT_MIN_SOURCES)
+        admin::get_min_sources_required(&env)
     }
 
     pub fn set_max_history_length(env: Env, new_max: u32) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        env.storage()
-            .persistent()
-            .set(&DataKey::MaxHistoryLength, &new_max);
-        MaxHistoryChangedEvent { value: new_max }.publish(&env);
+        admin::set_max_history_length(&env, new_max);
     }
 
     pub fn get_max_history_length(env: Env) -> u32 {
-        let key = DataKey::MaxHistoryLength;
-        if env.storage().persistent().has(&key) {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        env.storage()
-            .persistent()
-            .get(&key)
-            .unwrap_or(DEFAULT_MAX_HISTORY)
+        admin::get_max_history_length(&env)
     }
 
     pub fn set_resolution(env: Env, new_resolution: u32) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        env.storage()
-            .persistent()
-            .set(&DataKey::Resolution, &new_resolution);
-        ResolutionChangedEvent {
-            value: new_resolution,
-        }
-        .publish(&env);
+        admin::set_resolution(&env, new_resolution);
     }
 
     pub fn get_resolution(env: Env) -> u32 {
-        let key = DataKey::Resolution;
-        if env.storage().persistent().has(&key) {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        env.storage()
-            .persistent()
-            .get(&key)
-            .unwrap_or(DEFAULT_RESOLUTION)
+        admin::get_resolution(&env)
     }
 
     pub fn set_decimals(env: Env, new_decimals: u32) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        env.storage()
-            .persistent()
-            .set(&DataKey::Decimals, &new_decimals);
-        DecimalsChangedEvent {
-            value: new_decimals,
-        }
-        .publish(&env);
+        admin::set_decimals(&env, new_decimals);
     }
 
     pub fn get_decimals(env: Env) -> u32 {
-        let key = DataKey::Decimals;
-        if env.storage().persistent().has(&key) {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        env.storage()
-            .persistent()
-            .get(&key)
-            .unwrap_or(DEFAULT_DECIMALS)
+        admin::get_decimals(&env)
     }
 
     pub fn set_description(env: Env, new_description: String) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        env.storage()
-            .persistent()
-            .set(&DataKey::Description, &new_description);
-        DescriptionChangedEvent {
-            description: new_description.clone(),
-        }
-        .publish(&env);
+        admin::set_description(&env, new_description);
     }
 
     pub fn get_description(env: Env) -> String {
-        let key = DataKey::Description;
-        if env.storage().persistent().has(&key) {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        env.storage()
-            .persistent()
-            .get(&key)
-            .unwrap_or(String::from_str(&env, "Stellar Price Oracle"))
+        admin::get_description(&env)
     }
 
-    pub fn register_asset(env: Env, asset: Address) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        if env
-            .storage()
-            .persistent()
-            .has(&DataKey::AssetRegistered(asset.clone()))
-        {
-            panic_with_error!(env, ErrorCode::AssetAlreadyRegistered);
-        }
-        env.storage()
-            .persistent()
-            .set(&DataKey::AssetRegistered(asset.clone()), &true);
-        env.storage().persistent().set(
-            &DataKey::Aggregate(asset.clone()),
-            &AggregatePrice {
-                price: 0,
-                timestamp: 0,
-                num_sources: 0,
-                decimals: Self::get_decimals(env.clone()),
-            },
-        );
-        let mut assets = read_registered_assets(&env);
-        assets.push_back(asset.clone());
-        write_registered_assets(&env, &assets);
-        AssetRegisteredEvent {
-            asset: asset.clone(),
-        }
-        .publish(&env);
-    }
-
-    pub fn unregister_asset(env: Env, asset: Address) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        check_registered_asset(&env, &asset);
-        env.storage()
-            .persistent()
-            .remove(&DataKey::AssetRegistered(asset.clone()));
-        env.storage()
-            .persistent()
-            .remove(&DataKey::Aggregate(asset.clone()));
-        let assets = read_registered_assets(&env);
-        let mut new_assets: Vec<Address> = Vec::new(&env);
-        for i in 0..assets.len() {
-            let a = assets.get_unchecked(i);
-            if a != asset {
-                new_assets.push_back(a);
-            }
-        }
-        write_registered_assets(&env, &new_assets);
-        AssetUnregisteredEvent {
-            asset: asset.clone(),
-        }
-        .publish(&env);
-    }
-
-    pub fn is_asset_registered(env: Env, asset: Address) -> bool {
-        let key = DataKey::AssetRegistered(asset);
-        let exists: bool = env.storage().persistent().get(&key).unwrap_or(false);
-        if exists {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        exists
-    }
+    // --- Sources ---
 
     pub fn add_source(env: Env, source: Address, name: String) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        if env
-            .storage()
-            .persistent()
-            .has(&DataKey::Source(source.clone()))
-        {
-            panic_with_error!(env, ErrorCode::SourceAlreadyExists);
-        }
-        env.storage()
-            .persistent()
-            .set(&DataKey::Source(source.clone()), &true);
-
-        let mut oracle_sources: OracleSources = read_oracle_sources(&env);
-        oracle_sources.sources.push_back(source.clone());
-        let source_name = name.clone();
-        oracle_sources.metadata.set(source.clone(), name);
-        env.storage()
-            .persistent()
-            .set(&DataKey::OracleSources, &oracle_sources);
-        SourceAddedEvent {
-            source: source.clone(),
-            name: source_name,
-        }
-        .publish(&env);
+        sources::add_source(&env, source, name);
     }
 
     pub fn remove_source(env: Env, source: Address) {
-        let admin = get_admin(&env);
-        admin.require_auth();
-        if !env
-            .storage()
-            .persistent()
-            .has(&DataKey::Source(source.clone()))
-        {
-            panic_with_error!(env, ErrorCode::SourceNotFound);
-        }
-        env.storage()
-            .persistent()
-            .remove(&DataKey::Source(source.clone()));
-
-        let mut oracle_sources: OracleSources = read_oracle_sources(&env);
-        let mut new_sources: Vec<Address> = Vec::new(&env);
-        for i in 0..oracle_sources.sources.len() {
-            let s = oracle_sources.sources.get_unchecked(i);
-            if s != source {
-                new_sources.push_back(s);
-            }
-        }
-        oracle_sources.sources = new_sources;
-        let removed_source = source.clone();
-        oracle_sources.metadata.remove(source);
-        env.storage()
-            .persistent()
-            .set(&DataKey::OracleSources, &oracle_sources);
-        SourceRemovedEvent {
-            source: removed_source,
-        }
-        .publish(&env);
+        sources::remove_source(&env, source);
     }
 
     pub fn is_source(env: Env, source: Address) -> bool {
-        let key = DataKey::Source(source.clone());
-        let exists: bool = env.storage().persistent().get(&key).unwrap_or(false);
-        if exists {
-            env.storage()
-                .persistent()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        exists
+        sources::is_source(&env, source)
     }
 
     pub fn get_oracle_sources(env: Env) -> OracleSources {
-        read_oracle_sources(&env)
+        sources::get_oracle_sources(&env)
     }
 
+    // --- Assets ---
+
+    pub fn register_asset(env: Env, asset: Address) {
+        assets::register_asset(&env, asset);
+    }
+
+    pub fn unregister_asset(env: Env, asset: Address) {
+        assets::unregister_asset(&env, asset);
+    }
+
+    pub fn is_asset_registered(env: Env, asset: Address) -> bool {
+        assets::is_asset_registered(&env, asset)
+    }
+
+    // --- Prices ---
+
     pub fn submit_price(env: Env, source: Address, asset: Address, price: i128, timestamp: u64) {
-        source.require_auth();
-        check_source(&env, &source);
-        check_registered_asset(&env, &asset);
-
-        if price <= 0 {
-            panic_with_error!(env, ErrorCode::InvalidPrice);
-        }
-
-        let decimals = Self::get_decimals(env.clone());
-
-        let entry = PriceEntry {
-            price,
-            timestamp,
-            source: source.clone(),
-            decimals,
-        };
-
-        env.storage()
-            .persistent()
-            .set(&DataKey::Submission(asset.clone(), source.clone()), &entry);
-
-        PriceSubmittedEvent {
-            asset: asset.clone(),
-            source: source.clone(),
-            price,
-            timestamp,
-        }
-        .publish(&env);
-
-        let min_required = Self::get_min_sources_required(env.clone());
-        let oracle_sources: OracleSources = read_oracle_sources(&env);
-        let total_sources = oracle_sources.sources.len();
-
-        let mut valid_prices: Vec<i128> = Vec::new(&env);
-        let mut latest_timestamp: u64 = 0;
-        let mut contributing_sources: u32 = 0;
-
-        for i in 0..total_sources {
-            let src = oracle_sources.sources.get_unchecked(i);
-            let sub_key = DataKey::Submission(asset.clone(), src);
-            let sub: Option<PriceEntry> = env.storage().persistent().get(&sub_key);
-            if let Some(entry_data) = sub {
-                env.storage()
-                    .persistent()
-                    .extend_ttl(&sub_key, LEDGER_THRESHOLD, LEDGER_BUMP);
-                if entry_data.timestamp > latest_timestamp {
-                    latest_timestamp = entry_data.timestamp;
-                }
-                valid_prices.push_back(entry_data.price);
-                contributing_sources += 1;
-            }
-        }
-
-        if contributing_sources >= min_required && !valid_prices.is_empty() {
-            let median_price = compute_median(&valid_prices);
-
-            let current_ledger = env.ledger().sequence();
-            let agg_key = DataKey::Aggregate(asset.clone());
-            env.storage()
-                .persistent()
-                .extend_ttl(&agg_key, LEDGER_THRESHOLD, LEDGER_BUMP);
-            let prev_aggregate: AggregatePrice = env.storage().persistent().get(&agg_key).unwrap();
-
-            let aggregate = AggregatePrice {
-                price: median_price,
-                timestamp: latest_timestamp,
-                num_sources: contributing_sources,
-                decimals,
-            };
-            env.storage()
-                .persistent()
-                .set(&DataKey::Aggregate(asset.clone()), &aggregate);
-
-            if prev_aggregate.price != median_price || prev_aggregate.timestamp != latest_timestamp
-            {
-                let history_entry = PriceHistoryEntry {
-                    price: median_price,
-                    timestamp: latest_timestamp,
-                    ledger: current_ledger,
-                    num_sources: contributing_sources,
-                };
-                env.storage().temporary().set(
-                    &DataKey::PriceHistory(asset.clone(), current_ledger),
-                    &history_entry,
-                );
-            }
-
-            PriceAggregatedEvent {
-                asset: asset.clone(),
-                price: median_price,
-                num_sources: contributing_sources,
-                timestamp: latest_timestamp,
-            }
-            .publish(&env);
-        }
+        prices::submit_price(&env, source, asset, price, timestamp);
     }
 
     pub fn get_price(env: Env, asset: Address) -> AggregatePrice {
-        check_registered_asset(&env, &asset);
-        let key = DataKey::Aggregate(asset);
-        env.storage()
-            .persistent()
-            .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        let result: AggregatePrice = env.storage().persistent().get(&key).unwrap();
-        let resolution = Self::get_resolution(env.clone());
-        if resolution > 0 {
-            let ledger_time = env.ledger().timestamp();
-            if result.timestamp + (resolution as u64) < ledger_time {
-                panic_with_error!(env, ErrorCode::NoData);
-            }
-        }
-        result
+        prices::get_price(&env, asset)
     }
 
     pub fn get_source_price(env: Env, asset: Address, source: Address) -> PriceEntry {
-        check_registered_asset(&env, &asset);
-        check_source(&env, &source);
-        let key = DataKey::Submission(asset, source);
-        env.storage()
-            .persistent()
-            .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        env.storage().persistent().get(&key).unwrap()
+        prices::get_source_price(&env, asset, source)
     }
 
     pub fn get_all_prices(env: Env, asset: Address) -> Vec<PriceEntry> {
-        check_registered_asset(&env, &asset);
-        let oracle_sources: OracleSources = read_oracle_sources(&env);
-        let mut prices: Vec<PriceEntry> = Vec::new(&env);
-        for i in 0..oracle_sources.sources.len() {
-            let src = oracle_sources.sources.get_unchecked(i);
-            let sub_key = DataKey::Submission(asset.clone(), src);
-            let sub: Option<PriceEntry> = env.storage().persistent().get(&sub_key);
-            if let Some(entry) = sub {
-                env.storage()
-                    .persistent()
-                    .extend_ttl(&sub_key, LEDGER_THRESHOLD, LEDGER_BUMP);
-                prices.push_back(entry);
-            }
-        }
-        prices
+        prices::get_all_prices(&env, asset)
     }
 
+    pub fn get_latest_ledger(env: Env) -> u32 {
+        env.ledger().sequence()
+    }
+
+    // --- History ---
+
     pub fn get_historical_price(env: Env, asset: Address, ledger: u32) -> PriceHistoryEntry {
-        check_registered_asset(&env, &asset);
-        let key = DataKey::PriceHistory(asset, ledger);
-        env.storage()
-            .temporary()
-            .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        env.storage().temporary().get(&key).unwrap()
+        history::get_historical_price(&env, asset, ledger)
     }
 
     pub fn has_historical_price(env: Env, asset: Address, ledger: u32) -> bool {
-        if !env
-            .storage()
-            .persistent()
-            .has(&DataKey::AssetRegistered(asset.clone()))
-        {
-            return false;
-        }
-        let key = DataKey::PriceHistory(asset, ledger);
-        let exists = env.storage().temporary().has(&key);
-        if exists {
-            env.storage()
-                .temporary()
-                .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        }
-        exists
+        history::has_historical_price(&env, asset, ledger)
     }
 
     pub fn get_historical_prices(
@@ -544,32 +161,10 @@ impl PriceOracleContract {
         start_ledger: u32,
         end_ledger: u32,
     ) -> Vec<PriceHistoryEntry> {
-        check_registered_asset(&env, &asset);
-        let max_range = Self::get_max_history_length(env.clone());
-        if end_ledger - start_ledger > max_range {
-            panic_with_error!(env, ErrorCode::NoData);
-        }
-        let mut entries: Vec<PriceHistoryEntry> = Vec::new(&env);
-        let mut ledger = start_ledger;
-        while ledger <= end_ledger {
-            let key = DataKey::PriceHistory(asset.clone(), ledger);
-            if env.storage().temporary().has(&key) {
-                let entry: PriceHistoryEntry = env.storage().temporary().get(&key).unwrap();
-                env.storage()
-                    .temporary()
-                    .extend_ttl(&key, LEDGER_THRESHOLD, LEDGER_BUMP);
-                entries.push_back(entry);
-            }
-            ledger += 1;
-        }
-        entries
+        history::get_historical_prices(&env, asset, start_ledger, end_ledger)
     }
 
-    pub fn get_latest_ledger(env: Env) -> u32 {
-        env.ledger().sequence()
-    }
-
-    // ---- SEP-40 Oracle Interface ----
+    // --- SEP-40 Oracle Interface ---
 
     pub fn base(env: Env) -> Asset {
         Asset::Other(Symbol::new(&env, "USD"))
@@ -585,134 +180,19 @@ impl PriceOracleContract {
     }
 
     pub fn resolution(env: Env) -> u32 {
-        Self::get_resolution(env)
+        admin::get_resolution(&env)
     }
 
     pub fn lastprice(env: Env, asset: Asset) -> Option<PriceData> {
-        let addr = match asset {
-            Asset::Stellar(a) => a,
-            Asset::Other(_) => return None,
-        };
-        let reg_key = DataKey::AssetRegistered(addr.clone());
-        if !env.storage().persistent().get(&reg_key).unwrap_or(false) {
-            return None;
-        }
-        let agg_key = DataKey::Aggregate(addr);
-        let result: AggregatePrice = env.storage().persistent().get(&agg_key)?;
-        let resolution = Self::get_resolution(env.clone());
-        if resolution > 0 {
-            let ledger_time = env.ledger().timestamp();
-            if result.timestamp + (resolution as u64) < ledger_time {
-                return None;
-            }
-        }
-        env.storage()
-            .persistent()
-            .extend_ttl(&agg_key, LEDGER_THRESHOLD, LEDGER_BUMP);
-        Some(PriceData {
-            price: result.price,
-            timestamp: result.timestamp,
-        })
+        prices::lastprice(&env, asset)
     }
 
     pub fn price(env: Env, asset: Asset, timestamp: u64) -> Option<PriceData> {
-        let addr = match asset {
-            Asset::Stellar(a) => a,
-            Asset::Other(_) => return None,
-        };
-        let reg_key = DataKey::AssetRegistered(addr.clone());
-        if !env.storage().persistent().get(&reg_key).unwrap_or(false) {
-            return None;
-        }
-        let agg_key = DataKey::Aggregate(addr.clone());
-        if let Some(agg) = env
-            .storage()
-            .persistent()
-            .get::<_, AggregatePrice>(&agg_key)
-        {
-            if agg.timestamp == timestamp {
-                return Some(PriceData {
-                    price: agg.price,
-                    timestamp: agg.timestamp,
-                });
-            }
-        }
-        let current_ledger = env.ledger().sequence();
-        let start = current_ledger.saturating_sub(1000);
-        let mut ledger = current_ledger;
-        loop {
-            let hist_key = DataKey::PriceHistory(addr.clone(), ledger);
-            if let Some(entry) = env
-                .storage()
-                .temporary()
-                .get::<_, PriceHistoryEntry>(&hist_key)
-            {
-                if entry.timestamp <= timestamp {
-                    return Some(PriceData {
-                        price: entry.price,
-                        timestamp: entry.timestamp,
-                    });
-                }
-            }
-            if ledger == start {
-                break;
-            }
-            ledger -= 1;
-        }
-        None
+        prices::price(&env, asset, timestamp)
     }
 
     pub fn prices(env: Env, asset: Asset, records: u32) -> Option<Vec<PriceData>> {
-        let addr = match asset {
-            Asset::Stellar(a) => a,
-            Asset::Other(_) => return None,
-        };
-        let reg_key = DataKey::AssetRegistered(addr.clone());
-        if !env.storage().persistent().get(&reg_key).unwrap_or(false) {
-            return None;
-        }
-        if records == 0 {
-            return Some(Vec::new(&env));
-        }
-        let mut result: Vec<PriceData> = Vec::new(&env);
-        let current_ledger = env.ledger().sequence();
-        let max_to_check = (records * 10).min(10000);
-        let start = current_ledger.saturating_sub(max_to_check);
-        let mut ledger = current_ledger;
-        loop {
-            let hist_key = DataKey::PriceHistory(addr.clone(), ledger);
-            if let Some(entry) = env
-                .storage()
-                .temporary()
-                .get::<_, PriceHistoryEntry>(&hist_key)
-            {
-                result.push_back(PriceData {
-                    price: entry.price,
-                    timestamp: entry.timestamp,
-                });
-                if result.len() >= records {
-                    break;
-                }
-            }
-            if ledger == start {
-                break;
-            }
-            ledger -= 1;
-        }
-        if result.is_empty() {
-            let agg_key = DataKey::Aggregate(addr);
-            if let Some(agg) = env
-                .storage()
-                .persistent()
-                .get::<_, AggregatePrice>(&agg_key)
-            {
-                result.push_back(PriceData {
-                    price: agg.price,
-                    timestamp: agg.timestamp,
-                });
-            }
-        }
-        Some(result)
+        prices::prices(&env, asset, records)
     }
 }
